requirements/common.txt
alembic==1.0.2
atomicwrites==1.2.1
bleach==2.0.0
blinker==1.4
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.4.0
dominate==2.3.4
elasticsearch==6.3.1
entrypoints==0.2.3
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-CKEditor==0.4.3
Flask-Dance==1.3.0
Flask-Bootstrap==3.3.7.1
flask-heroku==0.1.9
Flask-HTTPAuth==3.2.3
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.3.0
Flask-Moment==0.6.0
Flask-PageDown==0.2.2
Flask-SQLAlchemy==2.3.2
Flask-SSLify==0.1.5
Flask-WTF==0.14.2
gunicorn==19.7.1
heroku==0.1.4
html5lib==0.999999999
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
Mako==1.0.7
Markdown==3.0.1
MarkupSafe==1.0
more-itertools==4.3.0
nltk==3.3
Pillow==5.3.0
pluggy==0.8.0
py==1.7.0
pycparser==2.19
pyparsing==2.2.2
pytest==3.9.3
python-dateutil==2.7.5
python-dotenv==0.6.5
python-editor==1.0.3
pytz==2018.7
rauth==0.7.3
requests==2.20.1
six==1.11.0
SQLAlchemy==1.2.12
SQLAlchemy-Utils==0.33.11
stripe==2.16.0
svgwrite==1.2.0
textblob==0.15.1
toolz==0.9.0
Tree==0.2.4
twilio==6.26.0
typing==3.6.6
urllib3==1.24.1
visitor==0.1.3
waitress==1.1.0
webencodings==0.5.1
Werkzeug==0.14.1
WTForms==2.2.1
wtforms-html5==0.2.2
